requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
sqlalchemy==2.0.49
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.24
aiofiles==24.1.0
bcrypt==3.2.2
python-dotenv==1.1.0
pydantic[email]==2.12.5
email-validator==2.2.0
apscheduler==3.11.2
slowapi==0.1.9
pytest==8.3.4
httpx==0.28.1
